refactor(phase-5b): extract bindTopWindowLinkInterceptor to src/boot/link-interceptor.ts
Continues phase 5. desktop.ts is now 3,180 LOC, down from the 3,695
where this refactor started — six boot modules now live in
src/boot/ totalling 697 LOC of focused setup code.

- src/boot/link-interceptor.ts (167 LOC) — capture-phase click
  handler that intercepts /wp-admin/ anchors in the parent shell
  and routes them through manager.open() instead of letting the
  browser navigate the whole tab. Pure: every dependency comes
  through arguments or imports (INITIAL_ORIGIN, native-url-remap,
  geometry helpers, deriveWindowId).
- desktop.ts: drops the inlined definition, the local
  tryNativeUrlRemap import (now consumed inside link-interceptor),
  and the local findDockEntryForUrl import.

Remaining boot extractions (createSessionSaver + trackedFetch +
the wp.desktop.* facade assembly) need careful work because
they sit at the heart of init() and capture the WindowManager
instance.

All green: lint, tsc --noEmit, vitest (106 files / 936 tests).
